refactor: move application query
