+++ apps/ui/gql/gql-types.ts
@@ -5499,6 +5499,180 @@ export type ApplicationsQuery = {
   } | null;
 };
 
+export type ApplicationQueryVariables = Exact<{
+  id: Scalars["ID"]["input"];
+}>;
+
+export type ApplicationQuery = {
+  application?: {
+    id: string;
+    pk?: number | null;
+    status?: ApplicationStatusChoice | null;
+    lastModifiedDate: string;
+    applicantType?: ApplicantTypeChoice | null;
+    additionalInformation?: string | null;
+    applicationRound: {
+      id: string;
+      sentDate?: string | null;
+      notesWhenApplyingFi?: string | null;
+      notesWhenApplyingEn?: string | null;
+      notesWhenApplyingSv?: string | null;
+      pk?: number | null;
+      nameFi?: string | null;
+      nameSv?: string | null;
+      nameEn?: string | null;
+      applicationPeriodBegin: string;
+      applicationPeriodEnd: string;
+      reservationPeriodBegin: string;
+      reservationPeriodEnd: string;
+      status?: ApplicationRoundStatusChoice | null;
+      applicationsCount?: number | null;
+      reservationUnitCount?: number | null;
+      statusTimestamp?: string | null;
+      termsOfUse?: {
+        id: string;
+        pk?: string | null;
+        termsType: TermsType;
+        nameFi?: string | null;
+        nameEn?: string | null;
+        nameSv?: string | null;
+        textFi?: string | null;
+        textEn?: string | null;
+        textSv?: string | null;
+      } | null;
+      reservationUnits: Array<{
+        id: string;
+        pk?: number | null;
+        nameFi?: string | null;
+        nameSv?: string | null;
+        nameEn?: string | null;
+        minPersons?: number | null;
+        maxPersons?: number | null;
+        images: Array<{
+          id: string;
+          imageUrl?: string | null;
+          largeUrl?: string | null;
+          mediumUrl?: string | null;
+          smallUrl?: string | null;
+          imageType: ImageType;
+        }>;
+        unit?: {
+          id: string;
+          pk?: number | null;
+          nameFi?: string | null;
+          nameSv?: string | null;
+          nameEn?: string | null;
+        } | null;
+      }>;
+    };
+    applicationSections?: Array<{
+      id: string;
+      pk?: number | null;
+      name: string;
+      status?: ApplicationSectionStatusChoice | null;
+      reservationMaxDuration: number;
+      numPersons: number;
+      reservationsEndDate: string;
+      reservationsBeginDate: string;
+      appliedReservationsPerWeek: number;
+      reservationMinDuration: number;
+      suitableTimeRanges: Array<{
+        id: string;
+        pk?: number | null;
+        beginTime: string;
+        endTime: string;
+        dayOfTheWeek: Weekday;
+        priority: Priority;
+      }>;
+      purpose?: {
+        id: string;
+        pk?: number | null;
+        nameFi?: string | null;
+        nameSv?: string | null;
+        nameEn?: string | null;
+      } | null;
+      reservationUnitOptions: Array<{
+        id: string;
+        pk?: number | null;
+        preferredOrder: number;
+        reservationUnit: {
+          id: string;
+          pk?: number | null;
+          nameFi?: string | null;
+          nameEn?: string | null;
+          nameSv?: string | null;
+          unit?: {
+            id: string;
+            pk?: number | null;
+            nameFi?: string | null;
+            nameEn?: string | null;
+            nameSv?: string | null;
+          } | null;
+          applicationRoundTimeSlots: Array<{
+            id: string;
+            weekday: number;
+            closed: boolean;
+            reservableTimes?: Array<{
+              begin: string;
+              end: string;
+            } | null> | null;
+          }>;
+        };
+      }>;
+      ageGroup?: {
+        id: string;
+        pk?: number | null;
+        minimum: number;
+        maximum?: number | null;
+      } | null;
+    }> | null;
+    contactPerson?: {
+      id: string;
+      pk?: number | null;
+      firstName: string;
+      lastName: string;
+      email?: string | null;
+      phoneNumber?: string | null;
+    } | null;
+    organisation?: {
+      id: string;
+      pk?: number | null;
+      nameFi?: string | null;
+      identifier?: string | null;
+      organisationType: OrganizationTypeChoice;
+      coreBusinessFi?: string | null;
+      yearEstablished?: number | null;
+      address?: {
+        id: string;
+        pk?: number | null;
+        postCode: string;
+        streetAddressFi?: string | null;
+        cityFi?: string | null;
+      } | null;
+    } | null;
+    homeCity?: {
+      id: string;
+      pk?: number | null;
+      nameFi?: string | null;
+      nameEn?: string | null;
+      nameSv?: string | null;
+    } | null;
+    billingAddress?: {
+      id: string;
+      pk?: number | null;
+      postCode: string;
+      streetAddressFi?: string | null;
+      cityFi?: string | null;
+    } | null;
+    user?: {
+      id: string;
+      name?: string | null;
+      email: string;
+      pk?: number | null;
+    } | null;
+  } | null;
+};
+
 export type CreateApplicationMutationVariables = Exact<{
   input: ApplicationCreateMutationInput;
 }>;
@@ -7336,11 +7510,140 @@ export type ApplicationCommonFragment = {
   } | null;
 };
 
-export type ApplicationQueryVariables = Exact<{
+export type ReserveeNameFieldsFragment = {
+  reserveeFirstName?: string | null;
+  reserveeLastName?: string | null;
+  reserveeEmail?: string | null;
+  reserveePhone?: string | null;
+  reserveeType?: CustomerTypeChoice | null;
+  reserveeOrganisationName?: string | null;
+  reserveeId?: string | null;
+};
+
+export type ReserveeBillingFieldsFragment = {
+  reserveeId?: string | null;
+  reserveeIsUnregisteredAssociation?: boolean | null;
+  reserveeAddressStreet?: string | null;
+  reserveeAddressCity?: string | null;
+  reserveeAddressZip?: string | null;
+  billingFirstName?: string | null;
+  billingLastName?: string | null;
+  billingPhone?: string | null;
+  billingEmail?: string | null;
+  billingAddressStreet?: string | null;
+  billingAddressCity?: string | null;
+  billingAddressZip?: string | null;
+};
+
+export type TermsOfUseNameFieldsFragment = {
+  nameFi?: string | null;
+  nameEn?: string | null;
+  nameSv?: string | null;
+};
+
+export type TermsOfUseTextFieldsFragment = {
+  id: string;
+  textFi?: string | null;
+  textEn?: string | null;
+  textSv?: string | null;
+};
+
+export type TermsOfUseFieldsFragment = {
+  pk?: string | null;
+  termsType: TermsType;
+  nameFi?: string | null;
+  nameEn?: string | null;
+  nameSv?: string | null;
+  id: string;
+  textFi?: string | null;
+  textEn?: string | null;
+  textSv?: string | null;
+};
+
+export type PricingFieldsFragment = {
+  id: string;
+  begins: string;
+  priceUnit: PriceUnit;
+  lowestPrice: string;
+  highestPrice: string;
+  taxPercentage: { id: string; pk?: number | null; value: string };
+};
+
+export type ImageFragment = {
+  id: string;
+  imageUrl?: string | null;
+  largeUrl?: string | null;
+  mediumUrl?: string | null;
+  smallUrl?: string | null;
+  imageType: ImageType;
+};
+
+export type LocationFieldsFragment = {
+  id: string;
+  addressStreetFi?: string | null;
+  addressZip: string;
+  addressCityFi?: string | null;
+};
+
+export type LocationFieldsI18nFragment = {
+  addressStreetEn?: string | null;
+  addressStreetSv?: string | null;
+  addressCityEn?: string | null;
+  addressCitySv?: string | null;
+  id: string;
+  addressStreetFi?: string | null;
+  addressZip: string;
+  addressCityFi?: string | null;
+};
+
+export type BannerNotificationCommonFragment = {
+  id: string;
+  level: BannerNotificationLevel;
+  activeFrom?: string | null;
+  message: string;
+  messageEn?: string | null;
+  messageFi?: string | null;
+  messageSv?: string | null;
+};
+
+export type MetadataSetsFragment = {
+  id: string;
+  minPersons?: number | null;
+  maxPersons?: number | null;
+  metadataSet?: {
+    id: string;
+    requiredFields: Array<{ id: string; fieldName: string }>;
+    supportedFields: Array<{ id: string; fieldName: string }>;
+  } | null;
+};
+
+export type TermsOfUseQueryVariables = Exact<{
+  termsType?: InputMaybe<TermsType>;
+}>;
+
+export type TermsOfUseQuery = {
+  termsOfUse?: {
+    edges: Array<{
+      node?: {
+        id: string;
+        pk?: string | null;
+        termsType: TermsType;
+        nameFi?: string | null;
+        nameEn?: string | null;
+        nameSv?: string | null;
+        textFi?: string | null;
+        textEn?: string | null;
+        textSv?: string | null;
+      } | null;
+    } | null>;
+  } | null;
+};
+
+export type ApplicationViewQueryVariables = Exact<{
   id: Scalars["ID"]["input"];
 }>;
 
-export type ApplicationQuery = {
+export type ApplicationViewQuery = {
   application?: {
     id: string;
     pk?: number | null;
@@ -7351,9 +7654,6 @@ export type ApplicationQuery = {
     applicationRound: {
       id: string;
       sentDate?: string | null;
-      notesWhenApplyingFi?: string | null;
-      notesWhenApplyingEn?: string | null;
-      notesWhenApplyingSv?: string | null;
       pk?: number | null;
       nameFi?: string | null;
       nameSv?: string | null;
@@ -7507,135 +7807,6 @@ export type ApplicationQuery = {
       email: string;
       pk?: number | null;
     } | null;
-  } | null;
-};
-
-export type ReserveeNameFieldsFragment = {
-  reserveeFirstName?: string | null;
-  reserveeLastName?: string | null;
-  reserveeEmail?: string | null;
-  reserveePhone?: string | null;
-  reserveeType?: CustomerTypeChoice | null;
-  reserveeOrganisationName?: string | null;
-  reserveeId?: string | null;
-};
-
-export type ReserveeBillingFieldsFragment = {
-  reserveeId?: string | null;
-  reserveeIsUnregisteredAssociation?: boolean | null;
-  reserveeAddressStreet?: string | null;
-  reserveeAddressCity?: string | null;
-  reserveeAddressZip?: string | null;
-  billingFirstName?: string | null;
-  billingLastName?: string | null;
-  billingPhone?: string | null;
-  billingEmail?: string | null;
-  billingAddressStreet?: string | null;
-  billingAddressCity?: string | null;
-  billingAddressZip?: string | null;
-};
-
-export type TermsOfUseNameFieldsFragment = {
-  nameFi?: string | null;
-  nameEn?: string | null;
-  nameSv?: string | null;
-};
-
-export type TermsOfUseTextFieldsFragment = {
-  id: string;
-  textFi?: string | null;
-  textEn?: string | null;
-  textSv?: string | null;
-};
-
-export type TermsOfUseFieldsFragment = {
-  pk?: string | null;
-  termsType: TermsType;
-  nameFi?: string | null;
-  nameEn?: string | null;
-  nameSv?: string | null;
-  id: string;
-  textFi?: string | null;
-  textEn?: string | null;
-  textSv?: string | null;
-};
-
-export type PricingFieldsFragment = {
-  id: string;
-  begins: string;
-  priceUnit: PriceUnit;
-  lowestPrice: string;
-  highestPrice: string;
-  taxPercentage: { id: string; pk?: number | null; value: string };
-};
-
-export type ImageFragment = {
-  id: string;
-  imageUrl?: string | null;
-  largeUrl?: string | null;
-  mediumUrl?: string | null;
-  smallUrl?: string | null;
-  imageType: ImageType;
-};
-
-export type LocationFieldsFragment = {
-  id: string;
-  addressStreetFi?: string | null;
-  addressZip: string;
-  addressCityFi?: string | null;
-};
-
-export type LocationFieldsI18nFragment = {
-  addressStreetEn?: string | null;
-  addressStreetSv?: string | null;
-  addressCityEn?: string | null;
-  addressCitySv?: string | null;
-  id: string;
-  addressStreetFi?: string | null;
-  addressZip: string;
-  addressCityFi?: string | null;
-};
-
-export type BannerNotificationCommonFragment = {
-  id: string;
-  level: BannerNotificationLevel;
-  activeFrom?: string | null;
-  message: string;
-  messageEn?: string | null;
-  messageFi?: string | null;
-  messageSv?: string | null;
-};
-
-export type MetadataSetsFragment = {
-  id: string;
-  minPersons?: number | null;
-  maxPersons?: number | null;
-  metadataSet?: {
-    id: string;
-    requiredFields: Array<{ id: string; fieldName: string }>;
-    supportedFields: Array<{ id: string; fieldName: string }>;
-  } | null;
-};
-
-export type TermsOfUseQueryVariables = Exact<{
-  termsType?: InputMaybe<TermsType>;
-}>;
-
-export type TermsOfUseQuery = {
-  termsOfUse?: {
-    edges: Array<{
-      node?: {
-        id: string;
-        pk?: string | null;
-        termsType: TermsType;
-        nameFi?: string | null;
-        nameEn?: string | null;
-        nameSv?: string | null;
-        textFi?: string | null;
-        textEn?: string | null;
-        textSv?: string | null;
-      } | null;
-    } | null>;
   } | null;
 };
 
@@ -8525,6 +8696,99 @@ export type ApplicationsSuspenseQueryHookResult = ReturnType<
 export type ApplicationsQueryResult = Apollo.QueryResult<
   ApplicationsQuery,
   ApplicationsQueryVariables
+>;
+export const ApplicationDocument = gql`
+  query Application($id: ID!) {
+    application(id: $id) {
+      ...ApplicationCommon
+      applicationRound {
+        id
+        sentDate
+        notesWhenApplyingFi
+        notesWhenApplyingEn
+        notesWhenApplyingSv
+        termsOfUse {
+          id
+          ...TermsOfUseFields
+        }
+      }
+    }
+  }
+  ${ApplicationCommonFragmentDoc}
+  ${TermsOfUseFieldsFragmentDoc}
+`;
+
+/**
+ * __useApplicationQuery__
+ *
+ * To run a query within a React component, call `useApplicationQuery` and pass it any options that fit your needs.
+ * When your component renders, `useApplicationQuery` returns an object from Apollo Client that contains loading, error, and data properties
+ * you can use to render your UI.
+ *
+ * @param baseOptions options that will be passed into the query, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options;
+ *
+ * @example
+ * const { data, loading, error } = useApplicationQuery({
+ *   variables: {
+ *      id: // value for 'id'
+ *   },
+ * });
+ */
+export function useApplicationQuery(
+  baseOptions: Apollo.QueryHookOptions<
+    ApplicationQuery,
+    ApplicationQueryVariables
+  > &
+    (
+      | { variables: ApplicationQueryVariables; skip?: boolean }
+      | { skip: boolean }
+    )
+) {
+  const options = { ...defaultOptions, ...baseOptions };
+  return Apollo.useQuery<ApplicationQuery, ApplicationQueryVariables>(
+    ApplicationDocument,
+    options
+  );
+}
+export function useApplicationLazyQuery(
+  baseOptions?: Apollo.LazyQueryHookOptions<
+    ApplicationQuery,
+    ApplicationQueryVariables
+  >
+) {
+  const options = { ...defaultOptions, ...baseOptions };
+  return Apollo.useLazyQuery<ApplicationQuery, ApplicationQueryVariables>(
+    ApplicationDocument,
+    options
+  );
+}
+export function useApplicationSuspenseQuery(
+  baseOptions?:
+    | Apollo.SkipToken
+    | Apollo.SuspenseQueryHookOptions<
+        ApplicationQuery,
+        ApplicationQueryVariables
+      >
+) {
+  const options =
+    baseOptions === Apollo.skipToken
+      ? baseOptions
+      : { ...defaultOptions, ...baseOptions };
+  return Apollo.useSuspenseQuery<ApplicationQuery, ApplicationQueryVariables>(
+    ApplicationDocument,
+    options
+  );
+}
+export type ApplicationQueryHookResult = ReturnType<typeof useApplicationQuery>;
+export type ApplicationLazyQueryHookResult = ReturnType<
+  typeof useApplicationLazyQuery
+>;
+export type ApplicationSuspenseQueryHookResult = ReturnType<
+  typeof useApplicationSuspenseQuery
+>;
+export type ApplicationQueryResult = Apollo.QueryResult<
+  ApplicationQuery,
+  ApplicationQueryVariables
 >;
 export const CreateApplicationDocument = gql`
   mutation CreateApplication($input: ApplicationCreateMutationInput!) {
@@ -10563,99 +10827,6 @@ export type BannerNotificationsListQueryResult = Apollo.QueryResult<
   BannerNotificationsListQuery,
   BannerNotificationsListQueryVariables
 >;
-export const ApplicationDocument = gql`
-  query Application($id: ID!) {
-    application(id: $id) {
-      ...ApplicationCommon
-      applicationRound {
-        id
-        sentDate
-        notesWhenApplyingFi
-        notesWhenApplyingEn
-        notesWhenApplyingSv
-        termsOfUse {
-          id
-          ...TermsOfUseFields
-        }
-      }
-    }
-  }
-  ${ApplicationCommonFragmentDoc}
-  ${TermsOfUseFieldsFragmentDoc}
-`;
-
-/**
- * __useApplicationQuery__
- *
- * To run a query within a React component, call `useApplicationQuery` and pass it any options that fit your needs.
- * When your component renders, `useApplicationQuery` returns an object from Apollo Client that contains loading, error, and data properties
- * you can use to render your UI.
- *
- * @param baseOptions options that will be passed into the query, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options;
- *
- * @example
- * const { data, loading, error } = useApplicationQuery({
- *   variables: {
- *      id: // value for 'id'
- *   },
- * });
- */
-export function useApplicationQuery(
-  baseOptions: Apollo.QueryHookOptions<
-    ApplicationQuery,
-    ApplicationQueryVariables
-  > &
-    (
-      | { variables: ApplicationQueryVariables; skip?: boolean }
-      | { skip: boolean }
-    )
-) {
-  const options = { ...defaultOptions, ...baseOptions };
-  return Apollo.useQuery<ApplicationQuery, ApplicationQueryVariables>(
-    ApplicationDocument,
-    options
-  );
-}
-export function useApplicationLazyQuery(
-  baseOptions?: Apollo.LazyQueryHookOptions<
-    ApplicationQuery,
-    ApplicationQueryVariables
-  >
-) {
-  const options = { ...defaultOptions, ...baseOptions };
-  return Apollo.useLazyQuery<ApplicationQuery, ApplicationQueryVariables>(
-    ApplicationDocument,
-    options
-  );
-}
-export function useApplicationSuspenseQuery(
-  baseOptions?:
-    | Apollo.SkipToken
-    | Apollo.SuspenseQueryHookOptions<
-        ApplicationQuery,
-        ApplicationQueryVariables
-      >
-) {
-  const options =
-    baseOptions === Apollo.skipToken
-      ? baseOptions
-      : { ...defaultOptions, ...baseOptions };
-  return Apollo.useSuspenseQuery<ApplicationQuery, ApplicationQueryVariables>(
-    ApplicationDocument,
-    options
-  );
-}
-export type ApplicationQueryHookResult = ReturnType<typeof useApplicationQuery>;
-export type ApplicationLazyQueryHookResult = ReturnType<
-  typeof useApplicationLazyQuery
->;
-export type ApplicationSuspenseQueryHookResult = ReturnType<
-  typeof useApplicationSuspenseQuery
->;
-export type ApplicationQueryResult = Apollo.QueryResult<
-  ApplicationQuery,
-  ApplicationQueryVariables
->;
 export const TermsOfUseDocument = gql`
   query TermsOfUse($termsType: TermsType) {
     termsOfUse(termsType: $termsType) {
@@ -10734,4 +10905,96 @@ export type TermsOfUseSuspenseQueryHookResult = ReturnType<
 export type TermsOfUseQueryResult = Apollo.QueryResult<
   TermsOfUseQuery,
   TermsOfUseQueryVariables
+>;
+export const ApplicationViewDocument = gql`
+  query ApplicationView($id: ID!) {
+    application(id: $id) {
+      ...ApplicationCommon
+      applicationRound {
+        id
+        sentDate
+        termsOfUse {
+          id
+          ...TermsOfUseFields
+        }
+      }
+    }
+  }
+  ${ApplicationCommonFragmentDoc}
+  ${TermsOfUseFieldsFragmentDoc}
+`;
+
+/**
+ * __useApplicationViewQuery__
+ *
+ * To run a query within a React component, call `useApplicationViewQuery` and pass it any options that fit your needs.
+ * When your component renders, `useApplicationViewQuery` returns an object from Apollo Client that contains loading, error, and data properties
+ * you can use to render your UI.
+ *
+ * @param baseOptions options that will be passed into the query, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options;
+ *
+ * @example
+ * const { data, loading, error } = useApplicationViewQuery({
+ *   variables: {
+ *      id: // value for 'id'
+ *   },
+ * });
+ */
+export function useApplicationViewQuery(
+  baseOptions: Apollo.QueryHookOptions<
+    ApplicationViewQuery,
+    ApplicationViewQueryVariables
+  > &
+    (
+      | { variables: ApplicationViewQueryVariables; skip?: boolean }
+      | { skip: boolean }
+    )
+) {
+  const options = { ...defaultOptions, ...baseOptions };
+  return Apollo.useQuery<ApplicationViewQuery, ApplicationViewQueryVariables>(
+    ApplicationViewDocument,
+    options
+  );
+}
+export function useApplicationViewLazyQuery(
+  baseOptions?: Apollo.LazyQueryHookOptions<
+    ApplicationViewQuery,
+    ApplicationViewQueryVariables
+  >
+) {
+  const options = { ...defaultOptions, ...baseOptions };
+  return Apollo.useLazyQuery<
+    ApplicationViewQuery,
+    ApplicationViewQueryVariables
+  >(ApplicationViewDocument, options);
+}
+export function useApplicationViewSuspenseQuery(
+  baseOptions?:
+    | Apollo.SkipToken
+    | Apollo.SuspenseQueryHookOptions<
+        ApplicationViewQuery,
+        ApplicationViewQueryVariables
+      >
+) {
+  const options =
+    baseOptions === Apollo.skipToken
+      ? baseOptions
+      : { ...defaultOptions, ...baseOptions };
+  return Apollo.useSuspenseQuery<
+    ApplicationViewQuery,
+    ApplicationViewQueryVariables
+  >(ApplicationViewDocument, options);
+}
+export type ApplicationViewQueryHookResult = ReturnType<
+  typeof useApplicationViewQuery
+>;
+export type ApplicationViewLazyQueryHookResult = ReturnType<
+  typeof useApplicationViewLazyQuery
+>;
+export type ApplicationViewSuspenseQueryHookResult = ReturnType<
+  typeof useApplicationViewSuspenseQuery
+>;
+export type ApplicationViewQueryResult = Apollo.QueryResult<
+  ApplicationViewQuery,
+  ApplicationViewQueryVariables
 >;
